Fixed warning

+++ src/editor/combobox.cpp
@@ -140,7 +140,7 @@ Combobox::get_box_offset()
   //return -(rect.get_height() * current_item);
 
   // open to the top
-  return -(rect.get_height() * item_list.size());
+  return -(rect.get_height() * (int)item_list.size());
 }
 
 bool
